Modify menu items of the sidebar
