+++ components/BoardDialog.tsx
