+++ providers/convex-client-provider.tsx
@@ -3,7 +3,7 @@
 import { ClerkProvider, useAuth } from "@clerk/nextjs"
 import { ConvexProviderWithClerk } from "convex/react-clerk"
 import { AuthLoading, Authenticated, ConvexReactClient } from "convex/react"
-import { Loading } from "@/components/loading"
+import { Loading } from "@/components/Loading"
 
 interface ConvexProviderProps {
     children: React.ReactNode
